Fix Windows/MSVC pointer conversion warnings
Make GLES buffer offsets and Windows native handles pointer-width explicit, and compare crop bounds using their unsigned width.

+++ dali/internal/imaging/common/pixel-buffer-impl.cpp
@@ -446,7 +446,7 @@ PixelBufferPtr PixelBuffer::NewCrop(const PixelBuffer& inBuffer, uint16_t x, uin
     int copyBytes       = destStrideBytes;
 
     // Clamp crop to right edge
-    if(DALI_UNLIKELY(x + cropDimensions.GetWidth() > inBuffer.mWidth))
+    if(DALI_UNLIKELY(static_cast<uint32_t>(x) + cropDimensions.GetWidth() > inBuffer.mWidth))
     {
       copyBytes = (inBuffer.mWidth - x) * bytesPerPixel;
     }
